requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==26.1.0
blinker==1.9.0
ccxt==4.5.44
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
Flask==3.1.3
frozenlist==1.8.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
numpy==2.4.3
pandas==3.0.1
propcache==0.4.1
pycares==5.0.1
pycparser==3.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
requests==2.32.5
rich==14.3.3
setuptools==82.0.1
six==1.17.0
ta==0.11.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
yarl==1.23.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
